Added a render lock to stop concurrency from mashing up state in a shared model.
When using a shared model for validation, concurrency can cause state read and write to crash, this is because rendering sets flags on the structs to mark them as pre-rendered and this causes all kinds of problems.

This lock now keeps things orderly when running concurrent requests for validation that require schema rendering.

Signed-off-by: Dave Shanley <dave@quobix.com>