Add support for pluggable Regular Express Engines.
By default the underlying JSONSchema validator uses the standard Go
regular expression engine; this engine supports the RE2 syntax which
may result in failures during validation of OAS specifcations that
use non-RE2 compliant patterns.

The change introduces a configuration construct for each validator
that may be passed at time of creation to affect the way it operates.

Initially this construct allows for the regular expression engine to
be overriden by a user-supplied variant.

Closes #111